+++ src/components/shop/pricing.ts
@@ -1,0 +1,4 @@
+export const PRICING = {
+  round: { regular: 1500, student: 800 },
+  polo: { regular: 2000, student: 1000 },
+};
